// src/types/dbEnums.ts
export enum Role {
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER',
  EMPLOYEE = 'EMPLOYEE',
}

export enum OrgAuditEventType {
  INVITE_SENT = 'INVITE_SENT',
  INVITE_RESENT = 'INVITE_RESENT',
  INVITE_REVOKED = 'INVITE_REVOKED',
  INVITE_ACCEPTED = 'INVITE_ACCEPTED',
  ROLE_CHANGED = 'ROLE_CHANGED',
  MEMBER_DEACTIVATED = 'MEMBER_DEACTIVATED',
  MEMBER_REACTIVATED = 'MEMBER_REACTIVATED',
  ORG_UPDATED = 'ORG_UPDATED',
}

export enum TaskEventType {
  CREATED = 'CREATED',
  ASSIGNED = 'ASSIGNED',
  REASSIGNED = 'REASSIGNED',
  STATUS_CHANGED = 'STATUS_CHANGED',
  PHOTO_ADDED = 'PHOTO_ADDED',
  VENDOR_LINKED = 'VENDOR_LINKED',
  BRANDING_VENDOR_LINKED = 'BRANDING_VENDOR_LINKED',
  ARTWORK_UPLOADED = 'ARTWORK_UPLOADED',
  DISTRIBUTION_RECORDED = 'DISTRIBUTION_RECORDED',
  COMMENT_ADDED = 'COMMENT_ADDED',
}